requirements.txt
python-dotenv==1.0.1
anthropic==0.22.0
discord.py==2.3.2
aiohttp==3.9.3 